requirements.txt
groq
playwright
selectolax
python-dotenv